feat：核心功能完善

+++ shpes-server/src/main/java/com/shpes/common/api/IErrorCode.java
@@ -7,7 +7,7 @@ public interface IErrorCode {
     /**
      * 获取错误码
      */
-    int getCode();
+    long getCode();
 
     /**
      * 获取错误信息
